+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 * Add `RequestCode::SetConfig` for set config toml file.
 * Add more fields to `Configuration`.
 * Add `RequestCode::DeleteSamplingPolicy`
+* Add `RequestCode::InternalNetworkList`
 
 ### Fixed