+++ CHANGELOG.md
@@ -5,6 +5,14 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+* Upgrade minimum supported Rust version (MSRV) to 1.65.
+* Upgrade rustls to 0.21.
+* Upgrade quinn to 0.10.
+
 ## [0.6.1] - 2023-04-17
 
 ### Added
@@ -152,6 +160,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 * `send_frame` and `recv_frame` to send and receive length-delimited frames.
 
+[Unreleased]: https://github.com/petabi/oinq/compare/0.6.1...main
 [0.6.1]: https://github.com/petabi/oinq/compare/0.6.0...0.6.1
 [0.6.0]: https://github.com/petabi/oinq/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/petabi/oinq/compare/0.4.1...0.5.0